Decouple contextHelp from the GUI module to workaround a circular import
