+++ source/NVDAObjects/window/excel.py
@@ -597,7 +597,7 @@ class ElementsListDialog(browseMode.ElementsListDialog):
 
 
 class EditCommentDialog(
-		gui.ContextHelpMixin,
+		gui.contextHelp.ContextHelpMixin,
 		wx.TextEntryDialog,  # wxPython does not seem to call base class initializer, put last in MRO
 ):
 	helpId = "ExcelReportingComments"
