+++ source/gui/installerGui.py
@@ -19,7 +19,6 @@ from logHandler import log
 import gui
 from gui import guiHelper
 from gui.dpiScalingHelper import DpiScalingHelperMixinWithoutInit
-from .contextHelp import ContextHelpMixin
 import tones
 import systemUtils
 
@@ -125,7 +124,7 @@ def doSilentInstall(
 
 class InstallerDialog(
 		DpiScalingHelperMixinWithoutInit,
-		gui.ContextHelpMixin,
+		gui.contextHelp.ContextHelpMixin,
 		wx.Dialog,  # wxPython does not seem to call base class initializer, put last in MRO
 ):
 
@@ -268,7 +267,7 @@ class InstallerDialog(
 
 class InstallingOverNewerVersionDialog(
 		DpiScalingHelperMixinWithoutInit,
-		gui.ContextHelpMixin,
+		gui.contextHelp.ContextHelpMixin,
 		wx.Dialog,  # wxPython does not seem to call base class initializer, put last in MRO
 ):
 	
@@ -330,7 +329,7 @@ def showInstallGui():
 
 
 class PortableCreaterDialog(
-		ContextHelpMixin,
+		gui.contextHelp.ContextHelpMixin,
 		wx.Dialog,  # wxPython does not seem to call base class initializer, put last in MRO
 ):
